correct category create method set default value for parent id

+++ database/migrations/2020_03_17_185824_create_categories_table.php
@@ -15,7 +15,7 @@ class CreateCategoriesTable extends Migration
     {
         Schema::create('categories', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('parent_id')->default(0);
+            $table->unsignedBigInteger('parent_id')->default(0);
             $table->string('name');
             $table->timestamps();
         });
